commited again

+++ src/main/java/com/app/repo/CandidateSchemesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.model.CandidateSchemes;
 
-public interface CandidateSchemesRepository extends JpaRepository<CandidateSchemes,Integer>{
+public interface CandidateSchemesRepository extends JpaRepository<CandidateSchemes,Long>{
 
 }
